silan/requirements.txt
asgiref==3.5.2
attrs==22.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.3
django-allauth==0.51.0
django-environ==0.9.0
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.1
Markdown==3.4.1
MarkupSafe==2.1.1
oauthlib==3.2.2
packaging==21.3
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pyparsing==3.0.9
pyrsistent==0.19.2
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
swagger-spec-validator==3.0.3
typing_extensions==4.4.0
uritemplate==4.1.1
urllib3==1.26.12